requirements.txt
Flask==3.1.2
Flask-Babel==4.0.0
python-dotenv==1.0.0
pillow==12.1.0
gunicorn==23.0.0
requests==2.32.3
piexif==1.1.3
